+++ bar/backend.h
@@ -10,7 +10,7 @@ struct backend {
     void (*loop)(struct bar *bar,
                  void (*expose)(const struct bar *bar),
                  void (*on_mouse)(struct bar *bar, enum mouse_event event,
-                                  int x, int y));
+                                  enum mouse_button btn, int x, int y));
     void (*commit)(const struct bar *bar);
     void (*refresh)(const struct bar *bar);
     void (*set_cursor)(struct bar *bar, const char *cursor);