+++ particles/map.c
@@ -61,26 +61,29 @@ expose(const struct exposable *exposable, pixman_image_t *pix, int x, int y, int
 
 static void
 on_mouse(struct exposable *exposable, struct bar *bar, enum mouse_event event,
-         int x, int y)
+         enum mouse_button btn, int x, int y)
 {
     const struct particle *p = exposable->particle;
     const struct eprivate *e = exposable->private;
 
-    if (exposable->on_click != NULL) {
+    if ((event == ON_MOUSE_MOTION &&
+         exposable->particle->have_on_click_template) ||
+        exposable->on_click[btn] != NULL)
+    {
         /* We have our own handler */
-        exposable_default_on_mouse(exposable, bar, event, x, y);
+        exposable_default_on_mouse(exposable, bar, event, btn, x, y);
         return;
     }
 
     int px = p->left_margin;
     if (x >= px && x < px + e->exposable->width) {
         if (e->exposable->on_mouse != NULL)
-            e->exposable->on_mouse(e->exposable, bar, event, x - px, y);
+            e->exposable->on_mouse(e->exposable, bar, event, btn, x - px, y);
         return;
     }
 
     /* In the left- or right margin */
-    exposable_default_on_mouse(exposable, bar, event, x, y);
+    exposable_default_on_mouse(exposable, bar, event, btn, x, y);
 }
 
 static struct exposable *
@@ -119,15 +122,12 @@ instantiate(const struct particle *particle, const struct tag_set *tags)
 
     assert(e->exposable != NULL);
 
-    char *on_click = tags_expand_template(particle->on_click_template, tags);
-    struct exposable *exposable = exposable_common_new(particle, on_click);
+    struct exposable *exposable = exposable_common_new(particle, tags);
     exposable->private = e;
     exposable->destroy = &exposable_destroy;
     exposable->begin_expose = &begin_expose;
     exposable->expose = &expose;
     exposable->on_mouse = &on_mouse;
-
-    free(on_click);
     return exposable;
 }