+++ tag.c
@@ -533,3 +533,11 @@ tags_expand_template(const char *template, const struct tag_set *tags)
 
     return formatted.s;
 }
+
+void
+tags_expand_templates(char *expanded[], const char *template[], size_t nmemb,
+                      const struct tag_set *tags)
+{
+    for (size_t i = 0; i < nmemb; i++)
+        expanded[i] = tags_expand_template(template[i], tags);
+}